Refactor: 주문 생성과 개수 차감 로직 분리
- 기존 두 로직이 붙어있어서 데드락이 자주 발생하게 되었기 때문에 수량을 바꾸는(`=update`) 부분을 `StockService`로 분리해서 `StockResponse`를 반환하도록 하고, 주문 생성 시 반환 값을 사용하도록 변경함
- 테스트 코드에서 커넥션 풀의 최대 개수마다 재고가 1개씩 차감되던 부분이 동시에 커넥션을 얻어서 동시에 로직을 수행하고 나오게 돼서 발생하는 것으로 판단해서 `Thread.sleep`을 추가함

Related to: #27
