+++ src/main/java/com/example/ddakdaegi/domain/order/service/OrderService.java
@@ -2,18 +2,16 @@ package com.example.ddakdaegi.domain.order.service;
 
 import com.example.ddakdaegi.domain.member.entity.Member;
 import com.example.ddakdaegi.domain.member.repository.MemberRepository;
-import com.example.ddakdaegi.domain.order.dto.request.PromotionProductRequest;
 import com.example.ddakdaegi.domain.order.dto.response.OrderDetailResponse;
 import com.example.ddakdaegi.domain.order.dto.response.OrderResponse;
+import com.example.ddakdaegi.domain.order.dto.response.StockResponse;
 import com.example.ddakdaegi.domain.order.entity.Order;
 import com.example.ddakdaegi.domain.order.entity.OrderPromotionProduct;
 import com.example.ddakdaegi.domain.order.repository.OrderPromotionProductRepository;
 import com.example.ddakdaegi.domain.order.repository.OrderRepository;
 import com.example.ddakdaegi.domain.promotion.entity.PromotionProduct;
-import com.example.ddakdaegi.domain.promotion.repository.PromotionProductRepository;
 import com.example.ddakdaegi.global.common.dto.AuthUser;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -27,41 +25,23 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class OrderService {
 
-	private final PromotionProductRepository promotionProductRepository;
 	private final OrderRepository orderRepository;
 	private final OrderPromotionProductRepository orderPromotionProductRepository;
 	private final MemberRepository memberRepository;
 
 
 	@Transactional
-	public OrderResponse createOrder(AuthUser authUser, List<PromotionProductRequest> promotionProductRequests) {
-		List<Long> promotionProductIds = promotionProductRequests.stream()
-			.map(PromotionProductRequest::getPromotionProductId)
-			.collect(Collectors.toList());
+	public OrderResponse createOrder(AuthUser authUser, StockResponse stockResponse) {
 
-		Map<Long, Long> promotionProductIdToQuantityMap = promotionProductRequests.stream()
-			.collect(
-				Collectors.toMap(PromotionProductRequest::getPromotionProductId, PromotionProductRequest::getQuantity));
+		Member getMember = memberRepository.findById(authUser.getId())
+			.orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없음"));
 
-		List<PromotionProduct> promotionProducts = promotionProductRepository.findAllByIdIn(promotionProductIds);
-
-		long totalPrice = 0L;
-
-		for (PromotionProduct promotionProduct : promotionProducts) {
-			Long quantity = promotionProductIdToQuantityMap.get(promotionProduct.getId());
-			if (promotionProduct.getStock() < quantity) {
-				throw new RuntimeException("재고가 부족합니다.");
-			}
-			promotionProduct.decreaseStock(quantity);
-			totalPrice += quantity * promotionProduct.getPrice();
-		}
-
-		Order newOrder = Order.of(Member.fromAuthUser(authUser), totalPrice);
+		Order newOrder = Order.of(getMember, stockResponse.getTotalPrice());
 		orderRepository.save(newOrder);
 
-		List<OrderPromotionProduct> orderPromotionProducts = promotionProducts.stream()
+		List<OrderPromotionProduct> orderPromotionProducts = stockResponse.getPromotionProducts().stream()
 			.map(promotionProduct -> {
-				Long quantity = promotionProductIdToQuantityMap.get(promotionProduct.getId());
+				Long quantity = stockResponse.getPromotionProductIdToQuantityMap().get(promotionProduct.getId());
 				return OrderPromotionProduct.of(newOrder, promotionProduct, quantity);
 			}).collect(Collectors.toList());
